Split CT migration service from base service
